+++ src/pages/consultation/ConsultationFormPage.tsx
@@ -1,23 +1,12 @@
 'use client';
 
 import type React from 'react';
-import { useState, useEffect } from 'react';
-import { useParams, useNavigate } from 'react-router-dom';
 import { Calendar, ArrowLeft, FileText, User, Edit, Phone, Plus, Search } from 'react-feather';
 import {
-  ConsultationType,
   ConsultationStatus,
-  type UpdateConsultationReqDto,
-  ConsultationReqDto,
+  consultationTypeLabels,
+  consultationStatusLabels,
 } from '../../types/consultation';
-import type { CustomerResDto } from '../../types/customer';
-import {
-  getConsultationById,
-  createConsultation,
-  updateConsultation,
-} from '../../api/consultation';
-import { useToast } from '../../context/useToast';
-import { useAuth } from '../../context/useAuth';
 import DashboardLayout from '../../components/layout/DashboardLayout';
 import Button from '../../components/ui/Button';
 import Input from '../../components/ui/Input';
@@ -26,358 +15,34 @@ import Card from '../../components/ui/Card';
 import CustomerSelectionModal from '../../components/customers/CustomerSelectionModal';
 import Alert from '../../components/ui/Alert';
 import CustomerHistory from '../../components/consultation/CustomerHistory';
-
-// 상담 유형 표시 레이블 매핑
-const consultationTypeLabels: Record<ConsultationType, string> = {
-  [ConsultationType.PHONE]: '전화상담',
-  [ConsultationType.VISIT]: '방문상담',
-  [ConsultationType.EMAIL]: '이메일상담',
-  [ConsultationType.OTHER]: '기타',
-};
-
-// 상담 상태 표시 레이블 매핑
-const consultationStatusLabels: Record<ConsultationStatus, string> = {
-  [ConsultationStatus.RESERVED]: '예약됨',
-  [ConsultationStatus.COMPLETED]: '완료',
-  [ConsultationStatus.CANCELED]: '취소됨',
-};
+import useConsultationForm from '../../hooks/useConsultationForm';
+import { useNavigate } from 'react-router-dom';
 
 const ConsultationFormPage: React.FC = () => {
-  const { id } = useParams<{ id?: string }>();
   const navigate = useNavigate();
-  const { showToast } = useToast();
-  const { user } = useAuth();
-  const isEditMode = !!id;
-  console.log(isEditMode);
-
-  // 폼 데이터 상태
-  const [formData, setFormData] = useState<ConsultationReqDto>({
-    agentId: user?.id ? Number(user.id) : 0,
-    customerId: undefined,
-    consultationType: ConsultationType.PHONE,
-    content: '',
-    consultationDate: new Date().toISOString(),
-    status: ConsultationStatus.RESERVED,
-  });
-
-  // 초기 데이터 상태 (서버에서 받아온 원본 데이터)
-  const [initialData, setInitialData] = useState<ConsultationReqDto | null>(null);
-
-  const [loading, setLoading] = useState(isEditMode);
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [errors, setErrors] = useState<Record<string, string>>({});
-  const [isCustomerModalOpen, setIsCustomerModalOpen] = useState(false);
-  const [selectedCustomer, setSelectedCustomer] = useState<CustomerResDto | null>(null);
-  const [isNewCustomer, setIsNewCustomer] = useState(false);
-  const [newCustomerData, setNewCustomerData] = useState({
-    name: '',
-    contact: '',
-    email: '',
-  });
-  const [originalStatus, setOriginalStatus] = useState<ConsultationStatus>(
-    ConsultationStatus.RESERVED
-  );
-
-  useEffect(() => {
-    // 에이전트 ID 설정
-    if (user?.id) {
-      setFormData((prev) => ({
-        ...prev,
-        agentId: Number(user.id),
-      }));
-    }
-
-    if (isEditMode && id) {
-      const fetchConsultation = async () => {
-        setLoading(true);
-        try {
-          const response = await getConsultationById(Number(id));
-          console.log(response);
-          if (response.success && response.data) {
-            const consultation = response.data;
-
-            // 서버에서 받아온 데이터로 formData 설정
-            // 소문자 값을 대문자 Enum으로 변환
-            const consultationData = {
-              agentId: consultation.agentId,
-              customerId: consultation.customer.id,
-              // 소문자 값을 대문자 Enum으로 변환 (서버에서 소문자로 오는 경우 대비)
-              consultationType: consultation.consultationType.toUpperCase() as ConsultationType,
-              content: consultation.content || '',
-              consultationDate: consultation.consultationDate || new Date().toISOString(),
-              status: consultation.status.toUpperCase() as ConsultationStatus,
-            };
-
-            // formData와 initialData 모두 설정
-            setFormData(consultationData);
-            setInitialData(consultationData); // 초기 데이터 저장
-
-            // 원래 상태 저장 (상태 변경 제한을 위해)
-            setOriginalStatus(consultationData.status);
-
-            // 고객 정보가 있으면 설정
-            if (consultation.customer) {
-              setSelectedCustomer(consultation.customer);
-            }
-          } else {
-            showToast(response.error || '상담 정보를 불러오는데 실패했습니다.', 'error');
-            navigate('/consultations', { replace: true });
-          }
-        } catch (error) {
-          console.error('상담 정보를 불러오는 중 오류가 발생했습니다:', error);
-          showToast('상담 정보를 불러오는 중 오류가 발생했습니다.', 'error');
-          navigate('/consultations', { replace: true });
-        } finally {
-          setLoading(false);
-        }
-      };
-
-      fetchConsultation();
-    }
-  }, [id, isEditMode, navigate, showToast, user]);
-
-  const handleChange = (
-    e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement | HTMLTextAreaElement>
-  ) => {
-    const { name, value } = e.target;
-
-    // 상태 변경 제한 로직
-    if (name === 'status' && isEditMode) {
-      const newStatus = value as ConsultationStatus;
-
-      // RESERVED 상태에서만 COMPLETED 또는 CANCELED로 변경 가능
-      if (originalStatus !== ConsultationStatus.RESERVED && newStatus !== originalStatus) {
-        showToast('예약 상태에서만 상태를 변경할 수 있습니다.', 'error');
-        return;
-      }
-
-      // RESERVED 상태에서는 COMPLETED 또는 CANCELED로만 변경 가능
-      if (
-        originalStatus === ConsultationStatus.RESERVED &&
-        newStatus !== ConsultationStatus.RESERVED &&
-        newStatus !== ConsultationStatus.COMPLETED &&
-        newStatus !== ConsultationStatus.CANCELED
-      ) {
-        showToast('예약 상태에서는 완료 또는 취소 상태로만 변경할 수 있습니다.', 'error');
-        return;
-      }
-    }
-
-    // 상담일 수정 제한 로직
-    if (
-      name === 'consultationDate' &&
-      isEditMode &&
-      originalStatus !== ConsultationStatus.RESERVED
-    ) {
-      showToast('예약 상태의 상담만 일자를 수정할 수 있습니다.', 'error');
-      return;
-    }
-
-    // consultationType과 status는 Enum 타입으로 변환
-    if (name === 'consultationType') {
-      setFormData((prev) => ({ ...prev, [name]: value as ConsultationType }));
-    } else if (name === 'status') {
-      setFormData((prev) => ({ ...prev, [name]: value as ConsultationStatus }));
-    } else {
-      setFormData((prev) => ({ ...prev, [name]: value }));
-    }
-
-    // 입력 필드 변경 시 해당 필드의 오류 메시지 제거
-    if (errors[name]) {
-      setErrors((prev) => {
-        const newErrors = { ...prev };
-        delete newErrors[name];
-        return newErrors;
-      });
-    }
-  };
-
-  // 신규 고객 정보 입력 핸들러
-  const handleNewCustomerChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setNewCustomerData((prev) => ({ ...prev, [name]: value }));
-
-    // 입력 필드 변경 시 해당 필드의 오류 메시지 제거
-    if (errors[`newCustomer.${name}`]) {
-      setErrors((prev) => {
-        const newErrors = { ...prev };
-        delete newErrors[`newCustomer.${name}`];
-        return newErrors;
-      });
-    }
-  };
-
-  // 고객 선택 모달에서 고객 선택 시 호출되는 함수
-  const handleCustomerSelect = (customer: CustomerResDto) => {
-    setSelectedCustomer(customer);
-    setFormData((prev) => ({
-      ...prev,
-      customerId: customer.id,
-    }));
-    setIsNewCustomer(false);
-
-    // 고객 ID 관련 오류 메시지 제거
-    if (errors.customerId) {
-      setErrors((prev) => {
-        const newErrors = { ...prev };
-        delete newErrors.customerId;
-        return newErrors;
-      });
-    }
-
-    showToast('고객이 성공적으로 선택되었습니다.', 'success');
-  };
-
-  // 고객 변경 버튼 클릭 핸들러
-  const handleChangeCustomer = () => {
-    setIsCustomerModalOpen(true);
-  };
-
-  // 신규 고객 모드 전환
-  const handleToggleNewCustomer = () => {
-    setIsNewCustomer(!isNewCustomer);
-    if (isNewCustomer) {
-      setSelectedCustomer(null);
-    } else {
-      setSelectedCustomer(null);
-      // 신규 고객 데이터 초기화
-      setNewCustomerData({
-        name: '',
-        contact: '',
-        email: '',
-      });
-    }
-  };
-
-  const validateForm = () => {
-    const newErrors: Record<string, string> = {};
-
-    // 고객 정보 검증
-    if (!isNewCustomer && !formData.customerId) {
-      newErrors.customerId = '고객을 선택해주세요.';
-    }
-
-    // 신규 고객 정보 검증
-    if (isNewCustomer) {
-      if (!newCustomerData.contact) {
-        newErrors['newCustomer.contact'] = '연락처를 입력해주세요.';
-      }
-    }
-
-    if (!formData.consultationDate) {
-      newErrors.consultationDate = '상담 일자를 선택해주세요.';
-    }
-
-    if (!formData.content) {
-      newErrors.content = '상담 내용을 입력해주세요.';
-    }
-
-    setErrors(newErrors);
-    return Object.keys(newErrors).length === 0;
-  };
-
-  // 변경된 필드만 추출하는 함수
-  const getChangedFields = (): Partial<UpdateConsultationReqDto> => {
-    // 초기 데이터가 없으면 빈 객체 반환 (이 경우는 발생하지 않아야 함)
-    if (!initialData) return { agentId: formData.agentId };
-
-    // 변경된 필드를 담을 객체
-    const changedFields: Partial<UpdateConsultationReqDto> = {
-      // agentId는 항상 포함
-      agentId: formData.agentId,
-    };
-
-    // 각 필드 비교 및 변경된 필드만 추가
-    if (formData.consultationType !== initialData.consultationType) {
-      changedFields.consultationType = formData.consultationType;
-    }
-
-    if (formData.content !== initialData.content) {
-      // 빈 문자열이면 undefined로 설정
-      changedFields.content = formData.content === '' ? undefined : formData.content;
-    }
-
-    if (formData.consultationDate !== initialData.consultationDate) {
-      changedFields.consultationDate = formData.consultationDate;
-    }
-
-    if (formData.status !== initialData.status) {
-      changedFields.status = formData.status;
-    }
-
-    return changedFields;
-  };
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    if (!validateForm()) {
-      return;
-    }
-
-    setIsSubmitting(true);
-    try {
-      let response;
-
-      // 수정 모드
-      if (isEditMode && id) {
-        // 변경된 필드만 추출
-        const updateData = getChangedFields();
-
-        // 변경된 필드가 agentId만 있는지 확인 (실질적인 변경이 없는 경우)
-        const hasRealChanges = Object.keys(updateData).length > 1; // agentId 외에 다른 필드가 있는지
-
-        if (!hasRealChanges) {
-          showToast('변경된 내용이 없습니다.', 'info');
-          setIsSubmitting(false);
-          return;
-        }
-
-        // PATCH 요청 보내기
-        response = await updateConsultation(Number(id), updateData);
-      } else {
-        // 등록 모드 (기존 코드 유지)
-        const consultationData = {
-          ...formData,
-          // 신규 고객 정보 추가
-          newCustomer: isNewCustomer
-            ? {
-                name: newCustomerData.name || undefined, // 이름이 없으면 undefined로 설정
-                contact: newCustomerData.contact, // 연락처는 필수
-                email: newCustomerData.email || undefined,
-              }
-            : undefined,
-        };
-
-        response = await createConsultation(consultationData);
-      }
-
-      if (response.success) {
-        showToast(
-          isEditMode
-            ? '상담 정보가 성공적으로 수정되었습니다.'
-            : '상담이 성공적으로 등록되었습니다.',
-          'success'
-        );
-        navigate('/consultations');
-      } else {
-        showToast(
-          response.error ||
-            (isEditMode ? '상담 정보 수정에 실패했습니다.' : '상담 등록에 실패했습니다.'),
-          'error'
-        );
-      }
-    } catch (error) {
-      console.error('상담 정보를 저장하는 중 오류가 발생했습니다:', error);
-      showToast('상담 정보를 저장하는 중 오류가 발생했습니다.', 'error');
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  // 상담 상태에 따른 필드 비활성화 여부 확인
-  const isDateFieldDisabled = isEditMode && originalStatus !== ConsultationStatus.RESERVED;
-  const isStatusChangeRestricted = isEditMode && originalStatus !== ConsultationStatus.RESERVED;
+  const {
+    formData,
+    loading,
+    isSubmitting,
+    errors,
+    originalStatus,
+    isEditMode,
+    handleChange,
+    handleNewCustomerChange,
+    handleSubmit,
+    isDateFieldDisabled,
+    isStatusChangeRestricted,
+    isConsultationImmutable,
+    selectedCustomer,
+    isNewCustomer,
+    newCustomerData,
+    isCustomerModalOpen,
+    handleOpenCustomerModal,
+    handleCloseCustomerModal,
+    handleToggleNewCustomer,
+    handleSelectCustomer,
+    setIsNewCustomer,
+  } = useConsultationForm();
 
   if (loading) {
     return (
@@ -444,7 +109,7 @@ const ConsultationFormPage: React.FC = () => {
                             variant={isNewCustomer ? 'outline' : 'primary'}
                             size="sm"
                             onClick={() => {
-                              setIsCustomerModalOpen(true);
+                              handleOpenCustomerModal();
                               setIsNewCustomer(false);
                             }}
                             disabled={isEditMode} // 수정 모드에서는 고객 변경 불가
@@ -548,7 +213,7 @@ const ConsultationFormPage: React.FC = () => {
                             type="button"
                             variant="outline"
                             size="sm"
-                            onClick={handleChangeCustomer}
+                            onClick={handleOpenCustomerModal}
                             leftIcon={<Edit size={14} />}
                           >
                             고객 변경
@@ -584,6 +249,7 @@ const ConsultationFormPage: React.FC = () => {
                         value={formData.consultationType}
                         onChange={handleChange}
                         className="block w-full px-4 py-2 rounded-md border border-gray-300 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500"
+                        disabled={isConsultationImmutable}
                       >
                         {Object.entries(consultationTypeLabels).map(([type, label]) => (
                           <option key={type} value={type}>
@@ -591,6 +257,11 @@ const ConsultationFormPage: React.FC = () => {
                           </option>
                         ))}
                       </select>
+                      {isConsultationImmutable && (
+                        <p className="mt-1 text-sm text-gray-400">
+                          예약 상담만 상담 유형을 수정할 수 있습니다.
+                        </p>
+                      )}
                     </div>
 
                     {/* 상담 상태 */}
@@ -620,17 +291,19 @@ const ConsultationFormPage: React.FC = () => {
                         >
                           {consultationStatusLabels[ConsultationStatus.COMPLETED]}
                         </option>
-                        <option
-                          value={ConsultationStatus.CANCELED}
-                          disabled={
-                            isStatusChangeRestricted &&
-                            (originalStatus as ConsultationStatus) !==
-                              ConsultationStatus.RESERVED &&
-                            formData.status !== ConsultationStatus.CANCELED
-                          }
-                        >
-                          {consultationStatusLabels[ConsultationStatus.CANCELED]}
-                        </option>
+                        {isEditMode && (
+                          <option
+                            value={ConsultationStatus.CANCELED}
+                            disabled={
+                              isStatusChangeRestricted &&
+                              (originalStatus as ConsultationStatus) !==
+                                ConsultationStatus.RESERVED &&
+                              formData.status !== ConsultationStatus.CANCELED
+                            }
+                          >
+                            {consultationStatusLabels[ConsultationStatus.CANCELED]}
+                          </option>
+                        )}
                       </select>
                       {isStatusChangeRestricted && (
                         <p className="mt-1 text-xs text-gray-500">
@@ -663,8 +336,14 @@ const ConsultationFormPage: React.FC = () => {
                       leftIcon={<Calendar size={18} />}
                       required
                       disabled={isDateFieldDisabled}
+                      readOnly={isConsultationImmutable}
                     />
-                    {isDateFieldDisabled && (
+                    {isConsultationImmutable && (
+                      <p className="mt-1 text-sm text-gray-400">
+                        예약 상담만 상담일을 수정할 수 있습니다.
+                      </p>
+                    )}
+                    {!isConsultationImmutable && isDateFieldDisabled && (
                       <p className="mt-1 text-xs text-gray-500">
                         예약 상태의 상담만 일자를 수정할 수 있습니다.
                       </p>
@@ -674,16 +353,22 @@ const ConsultationFormPage: React.FC = () => {
                   {/* 상담 내용 */}
                   <div>
                     <label className="block text-sm font-medium text-gray-700 mb-1 text-left">
-                      상담 내용 <span className="text-red-500">*</span>
+                      상담 내용
                     </label>
                     <Textarea
                       name="content"
                       value={formData.content || ''}
                       onChange={handleChange}
-                      placeholder="상담 내용을 입력하세요"
+                      placeholder="상담 내용을 입력하세요 (선택)"
                       className="min-h-[100px]"
                       error={errors.content}
+                      readOnly={isConsultationImmutable}
                     />
+                    {isConsultationImmutable && (
+                      <p className="mt-1 text-sm text-gray-400">
+                        예약 상담만 내용을 상담 내용을 수정할 수 있습니다.
+                      </p>
+                    )}
                   </div>
                 </div>
               </Card>
@@ -722,8 +407,8 @@ const ConsultationFormPage: React.FC = () => {
       {/* 고객 선택 모달 */}
       <CustomerSelectionModal
         isOpen={isCustomerModalOpen}
-        onClose={() => setIsCustomerModalOpen(false)}
-        onSelectCustomer={handleCustomerSelect}
+        onClose={handleCloseCustomerModal}
+        onSelectCustomer={handleSelectCustomer}
         selectedCustomerId={selectedCustomer?.id || null}
       />
     </DashboardLayout>
