+++ src/types/consultation.ts
@@ -56,3 +56,18 @@ export interface UpdateConsultationReqDto {
   consultationDate?: string; // LocalDateTime을 string으로 표현
   status?: ConsultationStatus;
 }
+
+// 상담 유형 표시 레이블 매핑
+export const consultationTypeLabels: Record<ConsultationType, string> = {
+  [ConsultationType.PHONE]: '전화상담',
+  [ConsultationType.VISIT]: '방문상담',
+  [ConsultationType.EMAIL]: '이메일상담',
+  [ConsultationType.OTHER]: '기타',
+};
+
+// 상담 상태 표시 레이블 매핑
+export const consultationStatusLabels: Record<ConsultationStatus, string> = {
+  [ConsultationStatus.RESERVED]: '예약됨',
+  [ConsultationStatus.COMPLETED]: '완료',
+  [ConsultationStatus.CANCELED]: '취소됨',
+};
